+++ pineappl_cli/src/optimize.rs
@@ -1,6 +1,8 @@
 use super::helpers::{self, Subcommand};
 use anyhow::Result;
 use clap::{Parser, ValueHint};
+use pineappl::fk_table::{FkAssumptions, FkTable};
+use std::convert::TryFrom;
 use std::path::PathBuf;
 
 /// Optimizes the internal data structure to minimize memory usage.
@@ -12,13 +14,36 @@ pub struct Opts {
     /// Path to the optimized PineAPPL file.
     #[clap(parse(from_os_str), value_hint = ValueHint::FilePath)]
     output: PathBuf,
+    #[clap(
+        long = "fk-table",
+        parse(try_from_str = FkAssumptions::try_from),
+        possible_values = &[
+            "Nf6Ind",
+            "Nf6Sym",
+            "Nf5Ind",
+            "Nf5Sym",
+            "Nf4Ind",
+            "Nf4Sym",
+            "Nf3Ind",
+            "Nf3Sym",
+        ],
+        value_name = "ASSUMPTIONS",
+    )]
+    fk_table: Option<FkAssumptions>,
 }
 
 impl Subcommand for Opts {
     fn run(&self) -> Result<()> {
         let mut grid = helpers::read_grid(&self.input)?;
-        grid.optimize();
-        helpers::write_grid(&self.output, &grid)
+
+        if let Some(assumptions) = self.fk_table {
+            let mut fk_table = FkTable::try_from(grid)?;
+            fk_table.optimize(assumptions);
+            helpers::write_grid(&self.output, fk_table.grid())
+        } else {
+            grid.optimize();
+            helpers::write_grid(&self.output, &grid)
+        }
     }
 }
 
@@ -31,14 +56,16 @@ mod tests {
 Optimizes the internal data structure to minimize memory usage
 
 USAGE:
-    pineappl optimize <INPUT> <OUTPUT>
+    pineappl optimize [OPTIONS] <INPUT> <OUTPUT>
 
 ARGS:
     <INPUT>     Path to the input grid
     <OUTPUT>    Path to the optimized PineAPPL file
 
 OPTIONS:
-    -h, --help    Print help information
+        --fk-table <ASSUMPTIONS>    [possible values: Nf6Ind, Nf6Sym, Nf5Ind, Nf5Sym, Nf4Ind,
+                                    Nf4Sym, Nf3Ind, Nf3Sym]
+    -h, --help                      Print help information
 ";
 
     #[test]